organize cell line info card props
