+++ src/components/CellLineInfoCard/DiseaseCellLineInfoCard.tsx
@@ -1,8 +1,16 @@
 import React from "react";
-import CellLineInfoCardBase, { DiseaseProps } from "./CellLineInfoCardBase";
+import CellLineInfoCardLayout from "./CellLineInfoCardBase";
 import { getCloneSummary, formatCellLineId } from "../../utils";
+import { Clone, ParentLine } from "../../component-queries/types";
+import { CellLineInfoCardBaseProps } from "./types";
 
-export const DiseaseCellLineInfoCard: React.FC<DiseaseProps> = (props) => {
+interface DiseaseCellLineInfoCardProps extends CellLineInfoCardBaseProps {
+    parentalLine: ParentLine;
+    snp: string;
+    clones: Clone[];
+}
+
+export const DiseaseCellLineInfoCard: React.FC<DiseaseCellLineInfoCardProps> = (props) => {
     const infoRows = [
         { key: "1", label: "SNP", children: props.snp },
         { key: "2", label: "Gene Symbol", children: props.geneSymbol },
@@ -17,7 +25,7 @@ export const DiseaseCellLineInfoCard: React.FC<DiseaseProps> = (props) => {
     ];
 
     return (
-        <CellLineInfoCardBase
+        <CellLineInfoCardLayout
             {...props}
             infoRows={infoRows}
             cloneSummary={getCloneSummary(props.clones)}
