+++ src/components/CellLineInfoCard/types.ts
@@ -1,0 +1,16 @@
+import { ReactNode } from "react";
+
+export interface CellLineInfoCardBaseProps {
+    href: string;
+    certificateOfAnalysis: string;
+    healthCertificate: string;
+    orderLink: string;
+    cellLineId: number;
+    geneName: string;
+    geneSymbol: string;
+}
+export interface InfoTableRow {
+    key: React.Key;
+    label: string;
+    children: ReactNode;
+}
